do place details (including reviews), favorites UI/UX

+++ android/app/src/main/java/app/tourism/ui/screens/main/categories/categories/CategoriesScreen.kt
@@ -5,14 +5,11 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -27,7 +24,7 @@ import app.tourism.ui.common.special.PlacesItem
 
 @Composable
 fun CategoriesScreen(
-    onPlaceClick: (id: Int) -> Unit,
+    onPlaceClick: (id: Long) -> Unit,
     onMapClick: () -> Unit,
     categoriesVM: CategoriesViewModel = hiltViewModel()
 ) {
