+++ android/app/src/main/java/app/tourism/ui/screens/main/home/home/HomeScreen.kt
@@ -23,7 +23,6 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.ScaffoldDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -60,7 +59,7 @@ import app.tourism.ui.theme.getStarColor
 @Composable
 fun HomeScreen(
     onSearchClick: (String) -> Unit,
-    onPlaceClick: (id: Int) -> Unit,
+    onPlaceClick: (id: Long) -> Unit,
     onMapClick: () -> Unit,
     onCategoryClicked: () -> Unit,
     homeVM: HomeViewModel = hiltViewModel(),
